Refactor CLI emit helpers
